mlmanager/pyproject.toml
[project]
name = "mlmanager"
description = ""
version = "0.0.1"
authors = [
    { name = "Alessandro Albanesi", email = "alessandro.albanesi@studio.unibo.it" }
]
dependencies = [
    "flask",
    "scikit-learn",
    "pandas",
    "numpy",
    "pyaml",
    "pymongo",
    "icecream",
    "python-dotenv",
    "pytest",
    "requests"
]


[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
